requirements.txt
pygsheets
psycopg2-binary
numpy
requests
datetime
pandas
matplotlib
ipykernel
tabulate
impyla
openpyxl
pydrive
looker_sdk
snowflake-connector-python==2.7.6
databricks-sql-connector
pytesseract
tabula
fitz
google-auth
google-auth-oauthlib
google-api-python-client
python-dotenv
